findlib support, more robust interaction with toplevel for figures
added findlib support, no more fooling around dynloading stuff via an
eval_string call. No more uncontrolled output from findlib.

the mechanism for figures is a lot more robust: no interference with
ocaml printers, explicitly call a Scirep.Show.* function to render an
insert on the document. The value to be rendered is stored in a queue,
and the queue is flushed after each phrase.