pictures transmitted by stdout capture
graphics used to be transmitted by producing "insert" values that were
read in memory using a hack (and the Obj module). This commit removed
this mechanism and copied the one used in mdx: pictures are printed on
stdout as <img> HTML elements, captured by temporary stdout
modification, and interpreted when producing the final HTML.

In addition to simplifying the way things work, "normal" prints are
also part of the final document (just like for mdx).